Struct rusoto_dax::Cluster

pub struct Cluster {
    pub active_nodes: Option<i64>,
    pub cluster_arn: Option<String>,
    pub cluster_discovery_endpoint: Option<Endpoint>,
    pub cluster_name: Option<String>,
    pub description: Option<String>,
    pub iam_role_arn: Option<String>,
    pub node_ids_to_remove: Option<Vec<String>>,
    pub node_type: Option<String>,
    pub nodes: Option<Vec<Node>>,
    pub notification_configuration: Option<NotificationConfiguration>,
    pub parameter_group: Option<ParameterGroupStatus>,
    pub preferred_maintenance_window: Option<String>,
    pub security_groups: Option<Vec<SecurityGroupMembership>>,
    pub status: Option<String>,
    pub subnet_group: Option<String>,
    pub total_nodes: Option<i64>,
}

Contains all of the attributes of a specific DAX cluster.

Fields
active_nodes: Option<i64>

The number of nodes in the cluster that are active (i.e., capable of serving requests).

cluster_arn: Option<String>

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

cluster_discovery_endpoint: Option<Endpoint>

The configuration endpoint for this DAX cluster, consisting of a DNS name and a port number. Client applications can specify this endpoint, rather than an individual node endpoint, and allow the DAX client software to intelligently route requests and responses to nodes in the DAX cluster.

cluster_name: Option<String>

The name of the DAX cluster.

description: Option<String>

The description of the cluster.

iam_role_arn: Option<String>

A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

node_ids_to_remove: Option<Vec<String>>

A list of nodes to be removed from the cluster.

node_type: Option<String>

The node type for the nodes in the cluster. (All nodes in a DAX cluster are of the same type.)

nodes: Option<Vec<Node>>

A list of nodes that are currently in the cluster.

notification_configuration: Option<NotificationConfiguration>

Describes a notification topic and its status. Notification topics are used for publishing DAX events to subscribers using Amazon Simple Notification Service (SNS).

parameter_group: Option<ParameterGroupStatus>

The parameter group being used by nodes in the cluster.

preferred_maintenance_window: Option<String>

A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00. Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

security_groups: Option<Vec<SecurityGroupMembership>>

A list of security groups, and the status of each, for the nodes in the cluster.

status: Option<String>

The current status of the cluster.

subnet_group: Option<String>

The subnet group where the DAX cluster is running.

total_nodes: Option<i64>

The total number of nodes in the cluster.