Struct rusoto_dax::CreateClusterRequest

pub struct CreateClusterRequest {
    pub availability_zones: Option<Vec<String>>,
    pub cluster_name: String,
    pub description: Option<String>,
    pub iam_role_arn: String,
    pub node_type: String,
    pub notification_topic_arn: Option<String>,
    pub parameter_group_name: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub replication_factor: i64,
    pub security_group_ids: Option<Vec<String>>,
    pub subnet_group_name: Option<String>,
    pub tags: Option<Vec<Tag>>,
}

Fields
availability_zones: Option<Vec<String>>

The Availability Zones (AZs) in which the cluster nodes will be created. All nodes belonging to the cluster are placed in these Availability Zones. Use this parameter if you want to distribute the nodes across multiple AZs.

cluster_name: String

The cluster identifier. This parameter is stored as a lowercase string.

Constraints:

A name must contain from 1 to 20 alphanumeric characters or hyphens.

The first character must be a letter.

A name cannot end with a hyphen or contain two consecutive hyphens.

description: Option<String>

A description of the cluster.

iam_role_arn: String

A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

node_type: String

The compute and memory capacity of the nodes in the cluster.

notification_topic_arn: Option<String>

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications will be sent.

The Amazon SNS topic owner must be same as the DAX cluster owner.

parameter_group_name: Option<String>

The parameter group to be associated with the DAX cluster.

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which maintenance on the DAX cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:05:00-sun:09:00

If you don't specify a preferred maintenance window when you create or modify a cache cluster, DAX assigns a 60-minute maintenance window on a randomly selected day of the week.

replication_factor: i64

The number of nodes in the DAX cluster. A replication factor of 1 will create a single-node cluster, without any read replicas. For additional fault tolerance, you can create a multiple node cluster with one or more read replicas. To do this, set ReplicationFactor to 2 or more.

AWS recommends that you have at least two read replicas per cluster.

security_group_ids: Option<Vec<String>>

A list of security group IDs to be assigned to each node in the DAX cluster. (Each of the security group ID is system-generated.)

If this parameter is not specified, DAX assigns the default VPC security group to each node.

subnet_group_name: Option<String>

The name of the subnet group to be used for the replication group.

DAX clusters can only run in an Amazon VPC environment. All of the subnets that you specify in a subnet group must exist in the same VPC.

tags: Option<Vec<Tag>>

A set of tags to associate with the DAX cluster.