Struct rusoto_dax::DynamodbAcceleratorClient

pub struct DynamodbAcceleratorClient { /* fields omitted */ }

A client for the Amazon DAX API.

Methods

impl DynamodbAcceleratorClient

pub fn new(region: Region) -> DynamodbAcceleratorClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> DynamodbAcceleratorClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,