Trait rusoto_dax::DynamodbAccelerator

pub trait DynamodbAccelerator {
    fn create_cluster(
        &self, 
        input: CreateClusterRequest
    ) -> RusotoFuture<CreateClusterResponse, CreateClusterError>;

    fn create_parameter_group(
        &self, 
        input: CreateParameterGroupRequest
    ) -> RusotoFuture<CreateParameterGroupResponse, CreateParameterGroupError>;

    fn create_subnet_group(
        &self, 
        input: CreateSubnetGroupRequest
    ) -> RusotoFuture<CreateSubnetGroupResponse, CreateSubnetGroupError>;

    fn decrease_replication_factor(
        &self, 
        input: DecreaseReplicationFactorRequest
    ) -> RusotoFuture<DecreaseReplicationFactorResponse, DecreaseReplicationFactorError>;

    fn delete_cluster(
        &self, 
        input: DeleteClusterRequest
    ) -> RusotoFuture<DeleteClusterResponse, DeleteClusterError>;

    fn delete_parameter_group(
        &self, 
        input: DeleteParameterGroupRequest
    ) -> RusotoFuture<DeleteParameterGroupResponse, DeleteParameterGroupError>;

    fn delete_subnet_group(
        &self, 
        input: DeleteSubnetGroupRequest
    ) -> RusotoFuture<DeleteSubnetGroupResponse, DeleteSubnetGroupError>;

    fn describe_clusters(
        &self, 
        input: DescribeClustersRequest
    ) -> RusotoFuture<DescribeClustersResponse, DescribeClustersError>;

    fn describe_default_parameters(
        &self, 
        input: DescribeDefaultParametersRequest
    ) -> RusotoFuture<DescribeDefaultParametersResponse, DescribeDefaultParametersError>;

    fn describe_events(
        &self, 
        input: DescribeEventsRequest
    ) -> RusotoFuture<DescribeEventsResponse, DescribeEventsError>;

    fn describe_parameter_groups(
        &self, 
        input: DescribeParameterGroupsRequest
    ) -> RusotoFuture<DescribeParameterGroupsResponse, DescribeParameterGroupsError>;

    fn describe_parameters(
        &self, 
        input: DescribeParametersRequest
    ) -> RusotoFuture<DescribeParametersResponse, DescribeParametersError>;

    fn describe_subnet_groups(
        &self, 
        input: DescribeSubnetGroupsRequest
    ) -> RusotoFuture<DescribeSubnetGroupsResponse, DescribeSubnetGroupsError>;

    fn increase_replication_factor(
        &self, 
        input: IncreaseReplicationFactorRequest
    ) -> RusotoFuture<IncreaseReplicationFactorResponse, IncreaseReplicationFactorError>;

    fn list_tags(
        &self, 
        input: ListTagsRequest
    ) -> RusotoFuture<ListTagsResponse, ListTagsError>;

    fn reboot_node(
        &self, 
        input: RebootNodeRequest
    ) -> RusotoFuture<RebootNodeResponse, RebootNodeError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<TagResourceResponse, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<UntagResourceResponse, UntagResourceError>;

    fn update_cluster(
        &self, 
        input: UpdateClusterRequest
    ) -> RusotoFuture<UpdateClusterResponse, UpdateClusterError>;

    fn update_parameter_group(
        &self, 
        input: UpdateParameterGroupRequest
    ) -> RusotoFuture<UpdateParameterGroupResponse, UpdateParameterGroupError>;

    fn update_subnet_group(
        &self, 
        input: UpdateSubnetGroupRequest
    ) -> RusotoFuture<UpdateSubnetGroupResponse, UpdateSubnetGroupError>;
}

Trait representing the capabilities of the Amazon DAX API. Amazon DAX clients implement this trait.

Required methods

fn create_cluster(
&self,
input: CreateClusterRequest
) -> RusotoFuture<CreateClusterResponse, CreateClusterError>

Creates a DAX cluster. All nodes in the cluster run the same DAX caching software.

fn create_parameter_group(
&self,
input: CreateParameterGroupRequest
) -> RusotoFuture<CreateParameterGroupResponse, CreateParameterGroupError>

Creates a new parameter group. A parameter group is a collection of parameters that you apply to all of the nodes in a DAX cluster.

fn create_subnet_group(
&self,
input: CreateSubnetGroupRequest
) -> RusotoFuture<CreateSubnetGroupResponse, CreateSubnetGroupError>

Creates a new subnet group.

fn decrease_replication_factor(
&self,
input: DecreaseReplicationFactorRequest
) -> RusotoFuture<DecreaseReplicationFactorResponse, DecreaseReplicationFactorError>

Removes one or more nodes from a DAX cluster.

You cannot use DecreaseReplicationFactor to remove the last node in a DAX cluster. If you need to do this, use DeleteCluster instead.

fn delete_cluster(
&self,
input: DeleteClusterRequest
) -> RusotoFuture<DeleteClusterResponse, DeleteClusterError>

Deletes a previously provisioned DAX cluster. DeleteCluster deletes all associated nodes, node endpoints and the DAX cluster itself. When you receive a successful response from this action, DAX immediately begins deleting the cluster; you cannot cancel or revert this action.

fn delete_parameter_group(
&self,
input: DeleteParameterGroupRequest
) -> RusotoFuture<DeleteParameterGroupResponse, DeleteParameterGroupError>

Deletes the specified parameter group. You cannot delete a parameter group if it is associated with any DAX clusters.

fn delete_subnet_group(
&self,
input: DeleteSubnetGroupRequest
) -> RusotoFuture<DeleteSubnetGroupResponse, DeleteSubnetGroupError>

Deletes a subnet group.

You cannot delete a subnet group if it is associated with any DAX clusters.

fn describe_clusters(
&self,
input: DescribeClustersRequest
) -> RusotoFuture<DescribeClustersResponse, DescribeClustersError>

Returns information about all provisioned DAX clusters if no cluster identifier is specified, or about a specific DAX cluster if a cluster identifier is supplied.

If the cluster is in the CREATING state, only cluster level information will be displayed until all of the nodes are successfully provisioned.

If the cluster is in the DELETING state, only cluster level information will be displayed.

If nodes are currently being added to the DAX cluster, node endpoint information and creation time for the additional nodes will not be displayed until they are completely provisioned. When the DAX cluster state is available, the cluster is ready for use.

If nodes are currently being removed from the DAX cluster, no endpoint information for the removed nodes is displayed.

fn describe_default_parameters(
&self,
input: DescribeDefaultParametersRequest
) -> RusotoFuture<DescribeDefaultParametersResponse, DescribeDefaultParametersError>

Returns the default system parameter information for the DAX caching software.

fn describe_events(
&self,
input: DescribeEventsRequest
) -> RusotoFuture<DescribeEventsResponse, DescribeEventsError>

Returns events related to DAX clusters and parameter groups. You can obtain events specific to a particular DAX cluster or parameter group by providing the name as a parameter.

By default, only the events occurring within the last hour are returned; however, you can retrieve up to 14 days' worth of events if necessary.

fn describe_parameter_groups(
&self,
input: DescribeParameterGroupsRequest
) -> RusotoFuture<DescribeParameterGroupsResponse, DescribeParameterGroupsError>

Returns a list of parameter group descriptions. If a parameter group name is specified, the list will contain only the descriptions for that group.

fn describe_parameters(
&self,
input: DescribeParametersRequest
) -> RusotoFuture<DescribeParametersResponse, DescribeParametersError>

Returns the detailed parameter list for a particular parameter group.

fn describe_subnet_groups(
&self,
input: DescribeSubnetGroupsRequest
) -> RusotoFuture<DescribeSubnetGroupsResponse, DescribeSubnetGroupsError>

Returns a list of subnet group descriptions. If a subnet group name is specified, the list will contain only the description of that group.

fn increase_replication_factor(
&self,
input: IncreaseReplicationFactorRequest
) -> RusotoFuture<IncreaseReplicationFactorResponse, IncreaseReplicationFactorError>

Adds one or more nodes to a DAX cluster.

fn list_tags(
&self,
input: ListTagsRequest
) -> RusotoFuture<ListTagsResponse, ListTagsError>

List all of the tags for a DAX cluster. You can call ListTags up to 10 times per second, per account.

fn reboot_node(
&self,
input: RebootNodeRequest
) -> RusotoFuture<RebootNodeResponse, RebootNodeError>

Reboots a single node of a DAX cluster. The reboot action takes place as soon as possible. During the reboot, the node status is set to REBOOTING.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<TagResourceResponse, TagResourceError>

Associates a set of tags with a DAX resource. You can call TagResource up to 5 times per second, per account.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<UntagResourceResponse, UntagResourceError>

Removes the association of tags from a DAX resource. You can call UntagResource up to 5 times per second, per account.

fn update_cluster(
&self,
input: UpdateClusterRequest
) -> RusotoFuture<UpdateClusterResponse, UpdateClusterError>

Modifies the settings for a DAX cluster. You can use this action to change one or more cluster configuration parameters by specifying the parameters and the new values.

fn update_parameter_group(
&self,
input: UpdateParameterGroupRequest
) -> RusotoFuture<UpdateParameterGroupResponse, UpdateParameterGroupError>

Modifies the parameters of a parameter group. You can modify up to 20 parameters in a single request by submitting a list parameter name and value pairs.

fn update_subnet_group(
&self,
input: UpdateSubnetGroupRequest
) -> RusotoFuture<UpdateSubnetGroupResponse, UpdateSubnetGroupError>

Modifies an existing subnet group.

Loading content...