Crate rusoto_dax

DAX is a managed caching service engineered for Amazon DynamoDB. DAX dramatically speeds up database reads by caching frequently-accessed data from DynamoDB, so applications can access that data with sub-millisecond latency. You can create a DAX cluster easily, using the AWS Management Console. With a few simple modifications to your code, your application can begin taking advantage of the DAX cluster and realize significant improvements in read performance.

If you're using the service, you're probably looking for DynamodbAcceleratorClient and DynamodbAccelerator.

Structs

Cluster
Contains all of the attributes of a specific DAX cluster.

CreateClusterRequest

CreateClusterResponse

CreateParameterGroupRequest

CreateParameterGroupResponse

CreateSubnetGroupRequest

CreateSubnetGroupResponse

DecreaseReplicationFactorRequest

DecreaseReplicationFactorResponse

DeleteClusterRequest

DeleteClusterResponse

DeleteParameterGroupRequest

DeleteParameterGroupResponse

DeleteSubnetGroupRequest

DeleteSubnetGroupResponse

DescribeClustersRequest

DescribeClustersResponse

DescribeDefaultParametersRequest

DescribeDefaultParametersResponse

DescribeEventsRequest

DescribeEventsResponse

DescribeParameterGroupsRequest

DescribeParameterGroupsResponse

DescribeParametersRequest

DescribeParametersResponse

DescribeSubnetGroupsRequest

DescribeSubnetGroupsResponse

DynamodbAcceleratorClient
A client for the Amazon DAX API.

Endpoint
Represents the information required for client programs to connect to the configuration endpoint for a DAX cluster, or to an individual node within the cluster.

Event
Represents a single occurrence of something interesting within the system. Some examples of events are creating a DAX cluster, adding or removing a node, or rebooting a node.

IncreaseReplicationFactorRequest

IncreaseReplicationFactorResponse

ListTagsRequest

ListTagsResponse

Node
Represents an individual node within a DAX cluster.

NodeTypeSpecificValue
Represents a parameter value that is applicable to a particular node type.

NotificationConfiguration
Describes a notification topic and its status. Notification topics are used for publishing DAX events to subscribers using Amazon Simple Notification Service (SNS).

Parameter
Describes an individual setting that controls some aspect of DAX behavior.

ParameterGroup
A named set of parameters that are applied to all of the nodes in a DAX cluster.

ParameterGroupStatus
The status of a parameter group.

ParameterNameValue
An individual DAX parameter.

RebootNodeRequest

RebootNodeResponse

SecurityGroupMembership
An individual VPC security group and its status.

Subnet
Represents the subnet associated with a DAX cluster. This parameter refers to subnets defined in Amazon Virtual Private Cloud (Amazon VPC) and used with DAX.

SubnetGroup

Represents the output of one of the following actions:

CreateSubnetGroup

ModifySubnetGroup

Tag
A description of a tag. Every tag is a key-value pair. You can add up to 50 tags to a single DAX cluster.

AWS-assigned tag names and values are automatically assigned the aws: prefix, which the user cannot assign. AWS-assigned tag names do not count towards the tag limit of 50. User-assigned tag names have the prefix user:.

You cannot backdate the application of a tag.

TagResourceRequest

TagResourceResponse

UntagResourceRequest

UntagResourceResponse

UpdateClusterRequest

UpdateClusterResponse

UpdateParameterGroupRequest

UpdateParameterGroupResponse

UpdateSubnetGroupRequest

UpdateSubnetGroupResponse

Enums

CreateClusterError
Errors returned by CreateCluster

CreateParameterGroupError
Errors returned by CreateParameterGroup

CreateSubnetGroupError
Errors returned by CreateSubnetGroup

DecreaseReplicationFactorError
Errors returned by DecreaseReplicationFactor

DeleteClusterError
Errors returned by DeleteCluster

DeleteParameterGroupError
Errors returned by DeleteParameterGroup

DeleteSubnetGroupError
Errors returned by DeleteSubnetGroup

DescribeClustersError
Errors returned by DescribeClusters

DescribeDefaultParametersError
Errors returned by DescribeDefaultParameters

DescribeEventsError
Errors returned by DescribeEvents

DescribeParameterGroupsError
Errors returned by DescribeParameterGroups

DescribeParametersError
Errors returned by DescribeParameters

DescribeSubnetGroupsError
Errors returned by DescribeSubnetGroups

IncreaseReplicationFactorError
Errors returned by IncreaseReplicationFactor

ListTagsError
Errors returned by ListTags

RebootNodeError
Errors returned by RebootNode

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateClusterError
Errors returned by UpdateCluster

UpdateParameterGroupError
Errors returned by UpdateParameterGroup

UpdateSubnetGroupError
Errors returned by UpdateSubnetGroup

Traits

DynamodbAccelerator
Trait representing the capabilities of the Amazon DAX API. Amazon DAX clients implement this trait.